fix(backend-python): move config info from database.py to config.py
